+++ CHANGELOG.md
@@ -4,13 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.24.1-dev]
+## [0.25.0-dev]
 ### Added
 - Updated elixir_script mix compiler to support compiling elixir_script paths in dependencies
+- Add `Collectable` implementations
+- Updated `for` implementation to use Collectable
 
 ### Fixed
 - JS module functions not translated properly when imported
 - Update fs dependency to 2.12
+- Incorrect handling of function heads with guards
 
 ## [0.24.0] - 2017-01-15